style: Extension 정리
