+++ TrackUs/TrackUs/Sources/Root/MainTabView.swift
@@ -30,7 +30,7 @@ struct MainTabView: View {
                     .tabItem {
                         Image(.runIcon)
                             .renderingMode(.template)
-                        Text("러닝")
+                        Text(Tab.running.tabName)
                     }
                     .tag(Tab.running)
                     
@@ -38,7 +38,7 @@ struct MainTabView: View {
                         .tabItem {
                             Image(.chattingIcon)
                                 .renderingMode(.template)
-                            Text("채팅")
+                            Text(Tab.chat.tabName)
                         }
                         .tag(Tab.chat)
                     
@@ -46,7 +46,7 @@ struct MainTabView: View {
                         .tabItem {
                             Image(.reportIcon)
                                 .renderingMode(.template)
-                            Text("리포트")
+                            Text(Tab.report.tabName)
                         }
                         .tag(Tab.report)
                     
@@ -55,7 +55,7 @@ struct MainTabView: View {
                         .tabItem {
                             Image(.profileIcon)
                                 .renderingMode(.template)
-                            Text("프로필")
+                            Text(Tab.profile.tabName)
                         }
                         .tag(Tab.profile)
                 }
@@ -85,8 +85,8 @@ struct MainTabView: View {
                                     Capsule()
                                         .foregroundStyle(.caution)
                                 )
-                             //Text를 탭바 중간에서 오른쪽으로 2/4 위치에 배치
-                            .offset(x: geometry.size.width * 1.5 / 4, y: -30)
+                            //Text를 탭바 중간에서 오른쪽으로 2/4 위치에 배치
+                                .offset(x: geometry.size.width * 1.5 / 4, y: -30)
                         }
                     }
                 }
